Changed To Use Gitlab File Variable Type
With Gitlab one can have CI/CD variables with two different types: `file` and `variable`. When the variable type is set to `file` gitlab will create a random file, with the contents you specified in the `value` field, and set an environment variable, with the name of the variable being what was set in the `key` field, to the file's location.

Another way this could be used is as follows:
```
  script:
    - mkdir -p ~/.ssh
    - chmod 700 ~/.ssh
    - echo -e "Host *\n\tStrictHostKeyChecking no\n\n" > ~/.ssh/config
    - chmod 600 "$SSH_PRIVATE_KEY_STAGING"
    - scp -i "$SSH_PRIVATE_KEY_STAGING" -r deploy/app_staging/. $DEPLOY_STAGING_SSH_HOST:~/app
    - scp -i "$SSH_PRIVATE_KEY_STAGING" -r src/project/conf/. $DEPLOY_STAGING_SSH_HOST:~/app/conf
    - ssh -i "$SSH_PRIVATE_KEY_STAGING" $DEPLOY_STAGING_SSH_HOST 'chmod 700 ~/app/app.sh'
    - ssh -i "$SSH_PRIVATE_KEY_STAGING" $DEPLOY_STAGING_SSH_HOST 'cd ~/app && ./app.sh --file docker-compose-staging.yml up'
    - ssh -i "$SSH_PRIVATE_KEY_STAGING" $DEPLOY_STAGING_SSH_HOST 'cd ~/app && ./app.sh --file docker-compose-staging.yml update'
```
The propose change only shortens the file by 4 characters, and the example in this comment add a lot more. It is more useful to use it in a one off scenario. Either way here is the (minor) change.